+++ src/extension.ts
@@ -5,14 +5,14 @@ import * as vscode from 'vscode';
 import { AVDTreeView } from './ui/AVDTreeView';
 import { SDKPlatformsTreeView } from './ui/SDKPlatformsTreeView';
 import { SDKToolsTreeView } from './ui/SDKToolsTreeView';
-import { subscribe } from './ext_util';
+import { showYesNoQuickPick, subscribe } from './ext_util';
 import { Manager } from './core';
 
 
 export function activate(context: vscode.ExtensionContext) {
-
 	console.log("Loaded");
 	const manager = Manager.getInstance();
+
 	//avd manager
 	new AVDTreeView(context, manager);
 
@@ -23,12 +23,15 @@ export function activate(context: vscode.ExtensionContext) {
 
 	let sdkbin = manager.getConfig().sdkManager ?? "";
 	subscribe(context, [
+		vscode.commands.registerCommand('avdmanager.pkg-install', async (node) => manager.sdk.pkgInstall(node)),
+		vscode.commands.registerCommand('avdmanager.pkg-uninstall', async (node) => manager.sdk.pkgUnInstall(node)),
 		vscode.commands.registerCommand('avdmanager.pkg-accept-license', () => { manager.sdk.acceptLicnese(sdkbin); }),
 		vscode.commands.registerCommand('avdmanager.pkg-update-all', () => { manager.sdk.updateAllPkg(); })
 	]);
 
 
 }
+
 
 
 // this method is called when your extension is deactivated